+++ R/backend_clustermq.R
@@ -10,7 +10,8 @@ drake_backend_clustermq <- function(config) {
   }
   config$workers <- clustermq::workers(
     n_jobs = config$settings$jobs,
-    template = config$settings$template
+    template = config$settings$template,
+    log_worker = config$settings$log_worker
   )
   config$logger$disk("set common data")
   suppressWarnings(cmq_set_common_data(config))