+++ NEWS.md
@@ -10,6 +10,7 @@
 
 * Transition to updated `lifecycle` that does not require badges to be in `man/figures`.
 * Improve error message for empty dynamic grouping variables (#1308, @saadaslam).
+* Expose the `log_worker` argument of `clustermq::workers()` to `make()` and `drake_config()` (#1305, @billdenney, @mschubert).
 
 # Version 7.12.4